Fix/favorite issue (#22)

+++ src/screens/myFavoritess/useFavorites.ts
@@ -13,10 +13,12 @@ import {
 } from '../../utils';
 import { useAsyncResource, resourceCache } from 'use-async-resource';
 import type { FavoritesScreenNavigationProps } from './MyFavoritesScreen';
-import { useEffect } from 'react';
+import { useEffect, useRef } from 'react';
 import { getFavoriteFoodItems } from '../../utils/DataServiceHelper';
 import type { ParamList } from 'src/navigaitons';
 import { ShowToast } from '../../utils';
+import { Alert } from 'react-native';
+import type { SwipeToDeleteRef } from '../../components';
 
 export function useFavorites() {
   const services = useServices();
@@ -25,6 +27,7 @@ export function useFavorites() {
 
   const navigation = useNavigation<FavoritesScreenNavigationProps>();
   const route = useRoute<RouteProp<ParamList, 'FavoritesScreen'>>();
+  const swipeToDeleteRef = useRef<SwipeToDeleteRef>(null);
 
   const [favoriteFoodReader, getUpdatedFavoriteFood] = useAsyncResource(
     getFavoriteFoodItems,
@@ -42,9 +45,12 @@ export function useFavorites() {
   const meal = getMealLog(date, route.params.logToMeal);
 
   const onSaveFoodLogs = async (favoriteFoodItem: FavoriteFoodItem) => {
-    await services.dataService.saveFoodLog(
-      createFoodLogFromFavoriteFoodItem(favoriteFoodItem, meal, date)
+    const foodLog = createFoodLogFromFavoriteFoodItem(
+      favoriteFoodItem,
+      meal,
+      date
     );
+    await services.dataService.saveFoodLog(foodLog);
     ShowToast('Log added');
     navigation.pop(1);
     navigation.navigate('BottomNavigation', {
@@ -53,7 +59,22 @@ export function useFavorites() {
   };
 
   const onDeleteFavoritePress = async (uuid: string) => {
-    await services.dataService.deleteFavoriteFoodItem(uuid);
+    Alert.alert('Are you sure want to delete this from your log?', undefined, [
+      {
+        text: 'Cancel',
+        style: 'cancel',
+      },
+      {
+        text: 'Delete',
+        onPress: async () => {
+          await services.dataService.deleteFavoriteFoodItem(uuid);
+          resourceCache(getFavoriteFoodItems).clear();
+          getUpdatedFavoriteFood(services);
+          swipeToDeleteRef?.current?.closeSwipe();
+        },
+        style: 'destructive',
+      },
+    ]);
   };
   const onUpdateFavoritePress = async (favoriteFoodItem: FavoriteFoodItem) => {
     await services.dataService.saveFavoriteFoodItem(favoriteFoodItem);
@@ -79,9 +100,11 @@ export function useFavorites() {
   };
 
   return {
-    favoriteFoodLogs: favoriteFoodReader(),
     branding,
-    onSaveFoodLogs,
+    favoriteFoodLogs: favoriteFoodReader(),
     navigateToFavoriteFoodLogEditor,
+    onDeleteFavoritePress,
+    onSaveFoodLogs,
+    swipeToDeleteRef,
   };
 }
